pyproject.toml
[tool.poetry]
name = "bulletin-board"
version = "0.1.0"
description = ""
authors = ["Alexander Gedz <gedz.buh@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
black = "^24.10.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
shell = "^1.0.1"
isort = "^5.13.2"
django-cors-headers = "^4.5.0"
drf-yasg = "^1.21.7"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
setuptools = "^75.1.0"
pillow = "^10.4.0"
djoser = "^2.2.3"
coverage = "^7.6.3"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
